+++ pom.xml
@@ -32,7 +32,7 @@
     <module>smart-server</module>
     <module>smart-alluxio</module>
     <module>smart-rule</module>
-    <module>smart-zeppelin</module>
+    <!--<module>smart-zeppelin</module>-->
     <module>smart-agent</module>
     <module>smart-dist</module>
   </modules>
