+++ smart-agent/pom.xml
@@ -34,9 +34,9 @@
   <dependencies>
     <dependency>
       <groupId>org.smartdata</groupId>
-      <artifactId>smart-server</artifactId>
+      <artifactId>smart-engine</artifactId>
       <version>0.1-SNAPSHOT</version>
-      <scope>provided</scope>
+      <!--<scope>provided</scope>-->
     </dependency>
     <dependency>
       <groupId>com.typesafe.akka</groupId>
