+++ smart-engine/src/main/java/org/smartdata/server/cluster/ServerDaemon.java
@@ -15,14 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.smartdata.server.utils;
+package org.smartdata.server.cluster;
 
-public class Constants {
-  public static final long ONE_SECOND_IN_MILLIS = 1000L;
+public interface ServerDaemon {
 
-  public static final long ONE_MINUTE_IN_MILLIS = 60 * ONE_SECOND_IN_MILLIS;
-
-  public static final long ONE_HOUR_IN_MILLIS = 60 * ONE_MINUTE_IN_MILLIS;
-
-  public static final long ONE_DAY_IN_MILLIS = 24 * ONE_HOUR_IN_MILLIS;
+  void becomeActive();
 }
