--- smart-engine/src/main/java/org/smartdata/server/engine/cluster/ClusterMembershipListener.java
@@ -1,46 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.smartdata.server.engine.cluster;
-
-import com.hazelcast.core.MemberAttributeEvent;
-import com.hazelcast.core.MembershipEvent;
-import com.hazelcast.core.MembershipListener;
-import org.smartdata.server.utils.HazelcastUtil;
-
-public class ClusterMembershipListener implements MembershipListener {
-  private final ServerDaemon daemon;
-
-  public ClusterMembershipListener(ServerDaemon daemon) {
-    this.daemon = daemon;
-  }
-
-  @Override
-  public void memberAdded(MembershipEvent membershipEvent) {
-  }
-
-  @Override
-  public void memberRemoved(MembershipEvent membershipEvent) {
-    if (HazelcastUtil.isMaster(HazelcastInstanceProvider.getInstance())) {
-      this.daemon.becomeActive();
-    }
-  }
-
-  @Override
-  public void memberAttributeChanged(MemberAttributeEvent memberAttributeEvent) {
-  }
-}
