+++ smart-engine/src/main/java/org/smartdata/server/engine/cmdlet/CmdletExecutorService.java
@@ -17,8 +17,8 @@
  */
 package org.smartdata.server.engine.cmdlet;
 
-import org.smartdata.server.engine.cmdlet.CmdletFactory;
 import org.smartdata.server.engine.CmdletManager;
+import org.smartdata.server.engine.cmdlet.CmdletFactory;
 import org.smartdata.server.engine.cmdlet.message.LaunchCmdlet;
 
 public abstract class CmdletExecutorService {
