+++ smart-engine/src/main/java/org/smartdata/server/engine/cmdlet/LocalCmdletExecutorService.java
@@ -64,7 +64,7 @@ public class LocalCmdletExecutorService extends CmdletExecutorService implements
 
   @Override
   public void report(StatusMessage status) {
-    cmdletManager.updateStatue(status);
+    cmdletManager.updateStatus(status);
   }
 
   private class StatusFetchTask implements Runnable {
