+++ smart-engine/src/main/java/org/smartdata/server/engine/cmdlet/agent/AgentConstants.java
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,9 +15,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.smartdata.server.engine.cluster;
+package org.smartdata.server.engine.cmdlet.agent;
 
-public interface ServerDaemon {
+public class AgentConstants {
 
-  void becomeActive();
+  public static final String MASTER_ADDRESS = "smartdata.agent.master";
+  public static final String MASTER_ACTOR_SYSTEM_NAME = "AgentMaster";
+  public static final String MASTER_ACTOR_NAME = "master";
 }
