+++ smart-server/src/main/java/org/smartdata/server/SmartServerDaemon.java
@@ -21,10 +21,10 @@ import com.hazelcast.core.HazelcastInstance;
 import org.smartdata.SmartContext;
 import org.smartdata.conf.SmartConf;
 import org.smartdata.server.engine.CmdletManager;
-import org.smartdata.server.engine.cluster.ClusterMembershipListener;
-import org.smartdata.server.engine.cluster.HazelcastInstanceProvider;
-import org.smartdata.server.engine.cluster.HazelcastWorker;
-import org.smartdata.server.engine.cluster.ServerDaemon;
+import org.smartdata.server.cluster.ClusterMembershipListener;
+import org.smartdata.server.cluster.HazelcastInstanceProvider;
+import org.smartdata.server.cluster.HazelcastWorker;
+import org.smartdata.server.cluster.ServerDaemon;
 import org.smartdata.server.engine.cmdlet.CmdletFactory;
 import org.smartdata.server.utils.HazelcastUtil;
